# Replace linear_loc with ravel_multi_index, which is 3x faster