```
In [42]: %timeit np.ravel_multi_index(o.coords, o.shape)
11.2 µs ± 1.32 µs per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [43]: %timeit sparse._coo.common.linear_loc(o.coords, o.shape)
35.2 µs ± 2.84 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```