@eric-wieser 2020-03-03
Failure caused by numpy/numpy#15690