@eric-wieser 2020-03-03
Left the name the same, it makes the patch much smaller